refactor: rename to signup
